% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banknotes}
\alias{banknotes}
\title{Swiss Bank Notes}
\format{
a dataframe with the following variables: 
\describe{
 \item{Status}{the status of the banknote: genuine or counterfeit}
 \item{Length}{Length of bill (mm)}
 \item{Left}{Width of left edge (mm)}
 \item{Right}{Width of right edge (mm)}
 \item{Bottom}{Bottom margin width (mm)}
 \item{Top}{Top margin width (mm)}
 \item{Diagonal}{Length of diagonal (mm)}
}
}
\source{
Flury, B. and Riedwyl, H. (1988). Multivariate Statistics: A 
practical approach. London: Chapman & Hall, Tables 1.1 and 1.2, pp. 5-8.
}
\usage{
data(banknotes)
}
\description{
This data set contains six measurements on 100 genuine and 
100 fradulent old Swiss banknotes
}
\keyword{datasets}
