% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bark-package-deprecated.R, R/depr_bark.r,
%   R/depr_sim.Friedman1.R, R/depr_sim.Friedman2.R, R/depr_sim.Friedman3.R,
%   R/depr_simCircle.R
\name{bark-package-deprecated}
\alias{bark-package-deprecated}
\alias{bark_mat}
\alias{sim.Friedman1}
\alias{sim.Friedman2}
\alias{sim.Friedman3}
\alias{sim.Circle}
\title{Deprecated functions in package \pkg{bark}.}
\usage{
bark_mat(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  type = "se",
  classification = FALSE,
  keepevery = 100,
  nburn = 100,
  nkeep = 100,
  printevery = 1000,
  keeptrain = FALSE,
  fixed = list(),
  tune = list(lstep = 0.5, frequL = 0.2, dpow = 1, upow = 0, varphistep = 0.5, phistep =
    1),
  theta = list()
)

sim.Friedman1(n, sd = 1)

sim.Friedman2(n, sd = 125)

sim.Friedman3(n, sd = 0.1)

sim.Circle(n, dim = 5)
}
\value{
List of deprecated functions
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("<function>-deprecated")}.
}
\section{\code{bark_mat}}{
  Old version with matrix inputs used for testing;
}

\section{\code{sim.Friedman1}}{

For \code{sim.Friedman1}, use \code{\link{sim_Friedman1}}.
}

\section{\code{sim.Friedman2}}{

For \code{sim.Friedman2}, use \code{\link{sim_Friedman2}}.
}

\section{\code{sim.Friedman3}}{

For \code{sim.Friedman3}, use \code{\link{sim_Friedman3}}.
}

\section{\code{sim.Circle}}{

For \code{sim.Circle}, use \code{\link{sim_circle}}.
}

\seealso{
Other bark deprecated functions: 
\code{\link{bark-deprecated}},
\code{\link{sim.Circle-deprecated}},
\code{\link{sim.Friedman1-deprecated}},
\code{\link{sim.Friedman2-deprecated}},
\code{\link{sim.Friedman3-deprecated}}

Other bark functions: 
\code{\link{bark-package}},
\code{\link{bark}()},
\code{\link{sim_Friedman1}()},
\code{\link{sim_Friedman2}()},
\code{\link{sim_Friedman3}()},
\code{\link{sim_circle}()}
}
\concept{bark deprecated functions}
\concept{bark functions}
\keyword{internal}
